feat: QueryDsl 의존성 추가 및 News 목록, 페이징 작업중

+++ src/main/java/com/luckyvicky/common/vo/PageVO.java
@@ -23,7 +23,7 @@ public class PageVO {
     private Integer page;
     private Integer size;
     private Integer limit;
-    private Integer totalCount;
+    private Long totalCount;
 
     private Integer totalPages;
     private Integer startIndex;
